Implement calculating total value of inventory and refactor code
