+++ src/main/java/seedu/address/MainApp.java
@@ -13,21 +13,21 @@ import seedu.address.commons.core.Version;
 import seedu.address.commons.exceptions.DataConversionException;
 import seedu.address.commons.util.ConfigUtil;
 import seedu.address.commons.util.StringUtil;
-import seedu.address.experimental.model.Model;
-import seedu.address.experimental.model.ModelManager;
-import seedu.address.experimental.model.ReadOnlyReroll;
-import seedu.address.experimental.model.ReadOnlyUserPrefs;
-import seedu.address.experimental.model.Reroll;
-import seedu.address.experimental.model.UserPrefs;
-import seedu.address.experimental.model.util.SampleDataUtil;
-import seedu.address.experimental.storage.JsonRerollStorage;
-import seedu.address.experimental.storage.JsonUserPrefsStorage;
-import seedu.address.experimental.storage.RerollStorage;
-import seedu.address.experimental.storage.Storage;
-import seedu.address.experimental.storage.StorageManager;
-import seedu.address.experimental.storage.UserPrefsStorage;
 import seedu.address.logic.Logic;
 import seedu.address.logic.LogicManager;
+import seedu.address.model.Model;
+import seedu.address.model.ModelManager;
+import seedu.address.model.ReadOnlyReroll;
+import seedu.address.model.ReadOnlyUserPrefs;
+import seedu.address.model.Reroll;
+import seedu.address.model.UserPrefs;
+import seedu.address.model.util.SampleDataUtil;
+import seedu.address.storage.JsonRerollStorage;
+import seedu.address.storage.JsonUserPrefsStorage;
+import seedu.address.storage.RerollStorage;
+import seedu.address.storage.Storage;
+import seedu.address.storage.StorageManager;
+import seedu.address.storage.UserPrefsStorage;
 import seedu.address.ui.Ui;
 import seedu.address.ui.UiManager;
