+++ src/main/java/seedu/address/logic/Logic.java
@@ -4,10 +4,10 @@ import java.nio.file.Path;
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
-import seedu.address.experimental.model.ReadOnlyReroll;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.model.ReadOnlyReroll;
 import seedu.address.model.entity.Entity;
 
 /**
