+++ src/main/java/seedu/address/logic/commands/BackCommand.java
@@ -4,6 +4,7 @@ import static java.util.Objects.requireNonNull;
 
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.UiSwitchMode;
+import seedu.address.model.Model;
 
 /**
  * Edits the details of an existing person in the address book.
@@ -18,7 +19,7 @@ public class BackCommand extends Command {
     }
 
     @Override
-    public CommandResult execute(seedu.address.experimental.model.Model model) throws CommandException {
+    public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
         return new CommandResult(MESSAGE_SUCCESS, false, false, UiSwitchMode.LIST);
     }
