+++ src/main/java/seedu/address/logic/commands/Command.java
@@ -1,7 +1,7 @@
 package seedu.address.logic.commands;
 
-import seedu.address.experimental.model.Model;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.Model;
 
 /**
  * Represents a command with hidden internal logic and the ability to be executed.
