+++ src/main/java/seedu/address/logic/commands/ExitCommand.java
@@ -1,7 +1,7 @@
 package seedu.address.logic.commands;
 
-import seedu.address.experimental.model.Model;
 import seedu.address.logic.parser.UiSwitchMode;
+import seedu.address.model.Model;
 
 /**
  * Terminates the program.
