+++ src/main/java/seedu/address/logic/commands/FindCommand.java
@@ -3,7 +3,7 @@ package seedu.address.logic.commands;
 import static java.util.Objects.requireNonNull;
 
 import seedu.address.commons.core.Messages;
-import seedu.address.experimental.model.Model;
+import seedu.address.model.Model;
 import seedu.address.model.entity.NameContainsKeywordsPredicate;
 
 /**
