+++ src/main/java/seedu/address/logic/commands/HelpCommand.java
@@ -1,7 +1,7 @@
 package seedu.address.logic.commands;
 
-import seedu.address.experimental.model.Model;
 import seedu.address.logic.parser.UiSwitchMode;
+import seedu.address.model.Model;
 
 /**
  * Format full help instructions for every command for display.
