+++ src/main/java/seedu/address/logic/commands/TemplateCommand.java
@@ -5,8 +5,8 @@ import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 
 import java.util.NoSuchElementException;
 
-import seedu.address.experimental.model.Model;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.Model;
 import seedu.address.model.entity.Entity;
 import seedu.address.model.entity.Name;
