+++ src/main/java/seedu/address/model/Reroll.java
@@ -1,4 +1,4 @@
-package seedu.address.experimental.model;
+package seedu.address.model;
 
 import static java.util.Objects.requireNonNull;
