+++ src/main/java/seedu/address/storage/RerollStorage.java
@@ -1,14 +1,15 @@
-package seedu.address.experimental.storage;
+package seedu.address.storage;
 
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.Optional;
 
 import seedu.address.commons.exceptions.DataConversionException;
-import seedu.address.experimental.model.ReadOnlyReroll;
+import seedu.address.model.ReadOnlyReroll;
+import seedu.address.model.Reroll;
 
 /**
- * Represents a storage for {@link seedu.address.experimental.model.Reroll}.
+ * Represents a storage for {@link Reroll}.
  */
 public interface RerollStorage {
     /**
