+++ src/main/java/seedu/address/storage/Storage.java
@@ -1,13 +1,13 @@
-package seedu.address.experimental.storage;
+package seedu.address.storage;
 
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.Optional;
 
 import seedu.address.commons.exceptions.DataConversionException;
-import seedu.address.experimental.model.ReadOnlyReroll;
-import seedu.address.experimental.model.ReadOnlyUserPrefs;
-import seedu.address.experimental.model.UserPrefs;
+import seedu.address.model.ReadOnlyReroll;
+import seedu.address.model.ReadOnlyUserPrefs;
+import seedu.address.model.UserPrefs;
 
 /**
  * API of the Storage component
