+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -9,9 +9,9 @@ import java.util.Arrays;
 import java.util.List;
 
 import seedu.address.commons.core.index.Index;
-import seedu.address.experimental.model.Model;
-import seedu.address.experimental.model.Reroll;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.Model;
+import seedu.address.model.Reroll;
 import seedu.address.model.entity.Entity;
 import seedu.address.model.entity.NameContainsKeywordsPredicate;
