+++ src/test/java/seedu/address/logic/commands/ExitCommandTest.java
@@ -5,8 +5,8 @@ import static seedu.address.logic.commands.ExitCommand.MESSAGE_EXIT_ACKNOWLEDGEM
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.experimental.model.Model;
-import seedu.address.experimental.model.ModelManager;
+import seedu.address.model.Model;
+import seedu.address.model.ModelManager;
 
 public class ExitCommandTest {
     private Model model = new ModelManager();
