+++ src/test/java/seedu/address/logic/commands/HelpCommandTest.java
@@ -5,8 +5,8 @@ import static seedu.address.logic.commands.HelpCommand.SHOWING_HELP_MESSAGE;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.experimental.model.Model;
-import seedu.address.experimental.model.ModelManager;
+import seedu.address.model.Model;
+import seedu.address.model.ModelManager;
 
 public class HelpCommandTest {
     private Model model = new ModelManager();
