+++ src/test/java/seedu/address/logic/commands/MakeCommandIntegrationTest.java
@@ -7,9 +7,9 @@ import static seedu.address.testutil.TypicalEntities.getTypicalReroll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import seedu.address.experimental.model.Model;
-import seedu.address.experimental.model.ModelManager;
-import seedu.address.experimental.model.UserPrefs;
+import seedu.address.model.Model;
+import seedu.address.model.ModelManager;
+import seedu.address.model.UserPrefs;
 import seedu.address.model.entity.Entity;
 import seedu.address.testutil.EntityBuilder;
