+++ src/test/java/seedu/address/logic/commands/MakeCommandTest.java
@@ -15,11 +15,11 @@ import org.junit.jupiter.api.Test;
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
-import seedu.address.experimental.model.Model;
-import seedu.address.experimental.model.ReadOnlyReroll;
-import seedu.address.experimental.model.ReadOnlyUserPrefs;
-import seedu.address.experimental.model.Reroll;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.Model;
+import seedu.address.model.ReadOnlyReroll;
+import seedu.address.model.ReadOnlyUserPrefs;
+import seedu.address.model.Reroll;
 import seedu.address.model.entity.Character;
 import seedu.address.model.entity.Classification;
 import seedu.address.model.entity.Entity;
